% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{unset_seed}
\alias{unset_seed}
\title{Revert to a non-reproducible, entropy-seeded RNG (undo \code{\link[=set_seed]{set_seed()}}).}
\usage{
unset_seed()
}
\value{
\code{NULL}, invisibly.
}
\description{
Revert to a non-reproducible, entropy-seeded RNG (undo \code{\link[=set_seed]{set_seed()}}).
}
